refactor(server): extract voice route domain (#592)
Move the thin TTS/STT HTTP handlers out of server.py into
agentwire/routes/voice.py as VoiceRoutesMixin: api_voices,
api_voice_status, api_say, api_local_tts, api_answer, handle_transcribe,
handle_send (plus handle_transcribe's private _decode_audio_to_wav). The
TTS/STT engine (speak, _tts_*, _probe_shim, ensure_managed_*, audio
helpers) stays on the base class and resolves through the MRO.

Closes #592

Built by dotdev.dev
